Rubbish/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
paddlepaddle==2.5.0
paddlehub==2.4.0
opencv-python==4.8.1.78
Pillow==10.0.0
numpy==1.24.3